Minor polish
