+++ src/testWithSpringBoot_3_0/java/org/openrewrite/java/spring/batch/RemoveDefaultBatchConfigurerTest.java
@@ -41,18 +41,18 @@ class RemoveDefaultBatchConfigurerTest implements RewriteTest {
         // language=java
         rewriteRun(
           java(
-                """
-            import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
-            class Foo extends DefaultBatchConfigurer {
-                @Override
-                public void setDataSource(javax.sql.DataSource dataSource) {
-                    // Datasource ignored; this method and comment should be removed
-                }
-            }
-            """, """
-            class Foo {
-            }
-            """)
+            """
+              import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
+              class Foo extends DefaultBatchConfigurer {
+                  @Override
+                  public void setDataSource(javax.sql.DataSource dataSource) {
+                      // Datasource ignored; this method and comment should be removed
+                  }
+              }
+              """, """
+              class Foo {
+              }
+              """)
         );
     }
 
@@ -61,22 +61,22 @@ class RemoveDefaultBatchConfigurerTest implements RewriteTest {
         // language=java
         rewriteRun(
           java(
-                """
-            import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
-            class Foo extends DefaultBatchConfigurer {
-                Foo() {
-                    super();
-                }
-                        
-                @Override
-                public void setDataSource(javax.sql.DataSource dataSource) {
-                    super.setDataSource(dataSource);
-                }
-            }
-            """, """
-            class Foo {
-            }
-            """)
+            """
+              import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
+              class Foo extends DefaultBatchConfigurer {
+                  Foo() {
+                      super();
+                  }
+
+                  @Override
+                  public void setDataSource(javax.sql.DataSource dataSource) {
+                      super.setDataSource(dataSource);
+                  }
+              }
+              """, """
+              class Foo {
+              }
+              """)
         );
     }
 
@@ -85,23 +85,23 @@ class RemoveDefaultBatchConfigurerTest implements RewriteTest {
         // language=java
         rewriteRun(
           java(
-                """
-            import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
-            class Foo extends DefaultBatchConfigurer {
-                @Override
-                public void setDataSource(javax.sql.DataSource dataSource) {
-                    super.setDataSource(dataSource);
-                    System.out.println("Additional statements should ensure method is not removed");
-                }
-            }
-            """, """
-            class Foo {
-                /*~~(TODO Used to override a DefaultBatchConfigurer method; reconsider if still needed)~~>*/
-                public void setDataSource(javax.sql.DataSource dataSource) {
-                    System.out.println("Additional statements should ensure method is not removed");
-                }
-            }
-            """)
+            """
+              import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
+              class Foo extends DefaultBatchConfigurer {
+                  @Override
+                  public void setDataSource(javax.sql.DataSource dataSource) {
+                      super.setDataSource(dataSource);
+                      System.out.println("Additional statements should ensure method is not removed");
+                  }
+              }
+              """, """
+              class Foo {
+                  /*~~(TODO Used to override a DefaultBatchConfigurer method; reconsider if still needed)~~>*/
+                  public void setDataSource(javax.sql.DataSource dataSource) {
+                      System.out.println("Additional statements should ensure method is not removed");
+                  }
+              }
+              """)
         );
     }
 
@@ -110,35 +110,35 @@ class RemoveDefaultBatchConfigurerTest implements RewriteTest {
         // language=java
         rewriteRun(
           java(
-                """
-            package bar;
-            public interface Bar {
-                void baz();
-            }
-            """),
+            """
+              package bar;
+              public interface Bar {
+                  void baz();
+              }
+              """),
           java(
-                """
-            import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
-            class Foo extends DefaultBatchConfigurer implements bar.Bar {
-                @Override
-                public void setDataSource(javax.sql.DataSource dataSource) {
-                    // Datasource ignored; this method and comment should be removed
-                }
-                
-                @Override
-                public void baz() {
-                    // Comment only, still retained
-                }
-            }
-            """, """
-            class Foo implements bar.Bar {
+            """
+              import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
+              class Foo extends DefaultBatchConfigurer implements bar.Bar {
+                  @Override
+                  public void setDataSource(javax.sql.DataSource dataSource) {
+                      // Datasource ignored; this method and comment should be removed
+                  }
 
-                @Override
-                public void baz() {
-                    // Comment only, still retained
-                }
-            }
-            """)
+                  @Override
+                  public void baz() {
+                      // Comment only, still retained
+                  }
+              }
+              """, """
+              class Foo implements bar.Bar {
+
+                  @Override
+                  public void baz() {
+                      // Comment only, still retained
+                  }
+              }
+              """)
         );
     }
 
@@ -147,18 +147,18 @@ class RemoveDefaultBatchConfigurerTest implements RewriteTest {
         // language=java
         rewriteRun(
           java(
-                """
-            import org.springframework.batch.core.configuration.annotation.BatchConfigurer;
-            import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
-            class FooConfig  {
-                public BatchConfigurer bean(javax.sql.DataSource dataSource) {
-                    return new DefaultBatchConfigurer(dataSource);
-                }
-            }
-            """, """
-            class FooConfig {
-            }
-            """)
+            """
+              import org.springframework.batch.core.configuration.annotation.BatchConfigurer;
+              import org.springframework.batch.core.configuration.annotation.DefaultBatchConfigurer;
+              class FooConfig  {
+                  public BatchConfigurer bean(javax.sql.DataSource dataSource) {
+                      return new DefaultBatchConfigurer(dataSource);
+                  }
+              }
+              """, """
+              class FooConfig {
+              }
+              """)
         );
     }
 }
